Start renaming view modules to ui-*
This renames the simplified-theme module to simplified-ui-theme in
order to better demonstrate that it's a "view" module in the MVC
architecture diagram.

Affects: https://jira.nypl.org/browse/SIMPLY-2338
