+++ simplified-app-shared/src/main/java/org/nypl/simplified/app/catalog/CatalogFeedActivity.kt
@@ -73,7 +73,7 @@ import org.nypl.simplified.profiles.api.ProfileReadableType
 import org.nypl.simplified.profiles.controller.api.ProfileFeedRequest
 import org.nypl.simplified.profiles.controller.api.ProfilesControllerType
 import org.nypl.simplified.stack.ImmutableStack
-import org.nypl.simplified.theme.ThemeControl
+import org.nypl.simplified.ui.theme.ThemeControl
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.net.MalformedURLException
