+++ simplified-app-shared/src/main/java/org/nypl/simplified/app/splash/SplashActivity.kt
@@ -32,7 +32,7 @@ import org.nypl.simplified.splash.SplashFragment
 import org.nypl.simplified.splash.SplashListenerType
 import org.nypl.simplified.splash.SplashParameters
 import org.nypl.simplified.taskrecorder.api.TaskResult
-import org.nypl.simplified.theme.ThemeControl
+import org.nypl.simplified.ui.theme.ThemeControl
 import org.nypl.simplified.threads.NamedThreadPools
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
