+++ simplified-app-vanilla-with-profiles/src/main/java/org/nypl/simplified/vanilla/with_profiles/BrandingThemeService.kt
@@ -1,7 +1,7 @@
 package org.nypl.simplified.vanilla.with_profiles
 
 import org.nypl.simplified.ui.branding.BrandingThemeOverrideServiceType
-import org.nypl.simplified.theme.ThemeValue
+import org.nypl.simplified.ui.theme.ThemeValue
 
 class BrandingThemeService : BrandingThemeOverrideServiceType {
   override fun overrideTheme(): ThemeValue {
