+++ simplified-tests-sandbox/build.gradle
@@ -7,7 +7,7 @@ dependencies {
   implementation project(":simplified-migration-fake")
   implementation project(":simplified-splash")
   implementation project(":simplified-tests-strings")
-  implementation project(":simplified-theme")
+  implementation project(":simplified-ui-theme")
 
   implementation libraries.androidXAppCompat
   implementation libraries.androidXConstraintLayout
