+++ simplified-tests/build.gradle
@@ -55,7 +55,7 @@ dependencies {
   api project(":simplified-stack")
   api project(":simplified-tenprint")
   api project(":simplified-tests-strings")
-  api project(":simplified-theme")
+  api project(":simplified-ui-theme")
   api project(":simplified-threads")
   api project(":simplified-ui-branding")
