+++ simplified-ui-branding/gradle.properties
@@ -1,4 +1,4 @@
-POM_ARTIFACT_ID=org.librarysimplified.branding
+POM_ARTIFACT_ID=org.librarysimplified.ui.branding
 POM_DESCRIPTION=Library Simplified (Branding functionality)
-POM_NAME=org.librarysimplified.branding
+POM_NAME=org.librarysimplified.ui.branding
 POM_PACKAGING=aar
