+++ simplified-ui-theme/gradle.properties
@@ -1,4 +1,4 @@
-POM_ARTIFACT_ID=org.librarysimplified.theme
+POM_ARTIFACT_ID=org.librarysimplified.ui.theme
 POM_DESCRIPTION=Library Simplified (Application theme functionality)
-POM_NAME=org.librarysimplified.theme
+POM_NAME=org.librarysimplified.ui.theme
 POM_PACKAGING=aar
