+++ simplified-ui-theme/src/main/res/drawable-hdpi/drawer_shadow.9.png
