+++ simplified-ui-theme/src/main/res/drawable-hdpi/ic_drawer.png
