+++ simplified-ui-theme/src/main/res/drawable-mdpi/drawer_shadow.9.png
