+++ simplified-ui-theme/src/main/res/drawable-mdpi/ic_drawer.png
