+++ simplified-ui-theme/src/main/res/drawable-xhdpi/drawer_shadow.9.png
