+++ simplified-ui-theme/src/main/res/drawable-xhdpi/ic_drawer.png
