+++ simplified-ui-theme/src/main/res/drawable-xxhdpi/drawer_shadow.9.png
