+++ simplified-ui-theme/src/main/res/drawable-xxhdpi/ic_drawer.png
