+++ simplified-ui-theme/src/main/res/drawable/drawer_shadow.9.png
