+++ simplified-ui-theme/src/main/res/drawable/ic_drawer.png
