+++ simplified-ui-theme/src/main/res/drawable/simplified_blank.xml
