+++ simplified-ui-theme/src/main/res/drawable/simplified_button_text.xml
