+++ simplified-ui-theme/src/main/res/layout/test.xml
