+++ simplified-ui-theme/src/main/res/values/styles.xml
