+++ simplified-ui-theme/src/main/sh/colors.txt
