+++ simplified-ui-theme/src/main/sh/styles.sh
